extract user orders to order_history
